completed until 4.23

+++ part4/blog/backend/controllers/blogs.js
@@ -1,8 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
@@ -15,7 +19,8 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const user = request.user
     const body = request.body
 
     if (!body.title || !body.url)
@@ -25,14 +30,28 @@ blogsRouter.post('/', async (request, response) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user.id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    // retrieve user from token
+    const user = request.user
+
+    const targetBlog = await Blog.findById(request.params.id)
+    if (!targetBlog)
+        response.status(404).end()
+
+    // do not allow operation if user.id !== blog.user
+    if(targetBlog.user.toString() !== user._id.toString())
+        return response.status(403).json({ 'error': 'forbidden request on resource' })
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
@@ -54,18 +73,6 @@ blogsRouter.put('/:id', async (request, response) => {
             { new: true }
         )
     response.status(200).json(updatedBlog)
-
-
-    // Blog
-    //     .findByIdAndUpdate(
-    //         request.params.id,
-    //         blog,
-    //         { new: true }
-    //     )
-    //     .then(updatedBlog => {
-    //         response.json(updatedBlog)
-    //     })
-    //     .catch(error => next(error))
 })
 
 module.exports = blogsRouter
